+++ cyxwiz-engine/src/gui/panels/gradient_descent_panel.cpp
@@ -249,7 +249,8 @@ void GradientDescentPanel::RenderStartPoint() {
 
 void GradientDescentPanel::RenderLoadingIndicator() {
     ImGui::Text("Computing optimization...");
-    float progress = 0.5f + 0.5f * std::sin(ImGui::GetTime() * 5.0f);
+    float progress = 0.5f + 0.5f * static_cast<float>(
+        std::sin(ImGui::GetTime() * 5.0));
     ImGui::ProgressBar(progress, ImVec2(-1, 0));
 }
