+++ cyxwiz-engine/src/gui/panels/integration_panel.cpp
@@ -230,7 +230,8 @@ void IntegrationPanel::RenderMethodSelector() {
 
 void IntegrationPanel::RenderLoadingIndicator() {
     ImGui::Text("Computing integral...");
-    float progress = 0.5f + 0.5f * std::sin(ImGui::GetTime() * 5.0f);
+    float progress = 0.5f + 0.5f * static_cast<float>(
+        std::sin(ImGui::GetTime() * 5.0));
     ImGui::ProgressBar(progress, ImVec2(-1, 0));
 }
