+++ cyxwiz-engine/src/gui/panels/lp_panel.cpp
@@ -209,7 +209,8 @@ void LPPanel::RenderConstraints() {
 
 void LPPanel::RenderLoadingIndicator() {
     ImGui::Text("Solving LP problem...");
-    float progress = 0.5f + 0.5f * std::sin(ImGui::GetTime() * 5.0f);
+    float progress = 0.5f + 0.5f * static_cast<float>(
+        std::sin(ImGui::GetTime() * 5.0));
     ImGui::ProgressBar(progress, ImVec2(-1, 0));
 }
